rpi/requirements.txt
influxdb
rpi-rfm69
sdnotify
